# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Token Gate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TokenInput {
  token: string;
}

export interface TokenVerifyResult {
  valid: boolean;
  secondsRemaining: number;
}

export interface TokenStatus {
  secondsRemaining: number;
  windowMinutes: number;
}

export interface AdminTokenResult {
  token: string;
  secondsRemaining: number;
}

export interface ErrorResult {
  error: string;
}

export interface ExamLink {
  label: string;
  description: string;
  url: string;
}

export type AppSettingsPrimaryColor = typeof AppSettingsPrimaryColor[keyof typeof AppSettingsPrimaryColor];


export const AppSettingsPrimaryColor = {
  blue: 'blue',
  indigo: 'indigo',
  emerald: 'emerald',
  rose: 'rose',
  slate: 'slate',
} as const;

export type AppSettingsBackgroundStyle = typeof AppSettingsBackgroundStyle[keyof typeof AppSettingsBackgroundStyle];


export const AppSettingsBackgroundStyle = {
  light: 'light',
  gray: 'gray',
  'blue-light': 'blue-light',
  warm: 'warm',
  gradient: 'gradient',
} as const;

export interface AppSettings {
  siteName: string;
  siteDescription: string;
  primaryColor: AppSettingsPrimaryColor;
  backgroundStyle: AppSettingsBackgroundStyle;
  examLinks: ExamLink[];
  /** Seconds of inactivity before redirect */
  inactivityTimeoutSeconds: number;
  /** How often the token rotates in minutes */
  tokenWindowMinutes: number;
  /** Whether a fixed custom token is active */
  useCustomToken: boolean;
}

export type AdminSettingsInputPrimaryColor = typeof AdminSettingsInputPrimaryColor[keyof typeof AdminSettingsInputPrimaryColor];


export const AdminSettingsInputPrimaryColor = {
  blue: 'blue',
  indigo: 'indigo',
  emerald: 'emerald',
  rose: 'rose',
  slate: 'slate',
} as const;

export type AdminSettingsInputBackgroundStyle = typeof AdminSettingsInputBackgroundStyle[keyof typeof AdminSettingsInputBackgroundStyle];


export const AdminSettingsInputBackgroundStyle = {
  light: 'light',
  gray: 'gray',
  'blue-light': 'blue-light',
  warm: 'warm',
  gradient: 'gradient',
} as const;

export interface AdminSettingsInput {
  siteName: string;
  siteDescription: string;
  primaryColor: AdminSettingsInputPrimaryColor;
  backgroundStyle: AdminSettingsInputBackgroundStyle;
  examLinks: ExamLink[];
  inactivityTimeoutSeconds: number;
  tokenWindowMinutes: number;
  useCustomToken: boolean;
  /** Fixed token value (only used when useCustomToken is true) */
  customToken?: string;
  /** New admin password (leave empty to keep current) */
  adminPassword?: string;
}