# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Token Gate API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: token
    description: Token operations
  - name: settings
    description: Site settings operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /token/verify:
    post:
      operationId: verifyToken
      tags: [token]
      summary: Verify a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenVerifyResult"

  /token/status:
    get:
      operationId: getTokenStatus
      tags: [token]
      summary: Get token rotation status
      responses:
        "200":
          description: Token rotation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenStatus"

  /token/admin:
    get:
      operationId: getAdminToken
      tags: [token]
      summary: Get current token (admin)
      parameters:
        - name: x-admin-secret
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTokenResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings (public)
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings (admin only)
      parameters:
        - name: x-admin-secret
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    TokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string

    TokenVerifyResult:
      type: object
      required: [valid, secondsRemaining]
      properties:
        valid:
          type: boolean
        secondsRemaining:
          type: integer

    TokenStatus:
      type: object
      required: [secondsRemaining, windowMinutes]
      properties:
        secondsRemaining:
          type: integer
        windowMinutes:
          type: integer

    AdminTokenResult:
      type: object
      required: [token, secondsRemaining]
      properties:
        token:
          type: string
        secondsRemaining:
          type: integer

    ErrorResult:
      type: object
      required: [error]
      properties:
        error:
          type: string

    ExamLink:
      type: object
      required: [label, description, url]
      properties:
        label:
          type: string
        description:
          type: string
        url:
          type: string

    AppSettings:
      type: object
      required:
        - siteName
        - siteDescription
        - primaryColor
        - backgroundStyle
        - examLinks
        - inactivityTimeoutSeconds
        - tokenWindowMinutes
        - useCustomToken
      properties:
        siteName:
          type: string
        siteDescription:
          type: string
        primaryColor:
          type: string
          enum: [blue, indigo, emerald, rose, slate]
        backgroundStyle:
          type: string
          enum: [light, gray, blue-light, warm, gradient]
        examLinks:
          type: array
          items:
            $ref: "#/components/schemas/ExamLink"
        inactivityTimeoutSeconds:
          type: integer
          description: Seconds of inactivity before redirect
        tokenWindowMinutes:
          type: integer
          description: How often the token rotates in minutes
        useCustomToken:
          type: boolean
          description: Whether a fixed custom token is active

    AdminSettingsInput:
      type: object
      required:
        - siteName
        - siteDescription
        - primaryColor
        - backgroundStyle
        - examLinks
        - inactivityTimeoutSeconds
        - tokenWindowMinutes
        - useCustomToken
      properties:
        siteName:
          type: string
        siteDescription:
          type: string
        primaryColor:
          type: string
          enum: [blue, indigo, emerald, rose, slate]
        backgroundStyle:
          type: string
          enum: [light, gray, blue-light, warm, gradient]
        examLinks:
          type: array
          items:
            $ref: "#/components/schemas/ExamLink"
        inactivityTimeoutSeconds:
          type: integer
        tokenWindowMinutes:
          type: integer
        useCustomToken:
          type: boolean
        customToken:
          type: string
          description: Fixed token value (only used when useCustomToken is true)
        adminPassword:
          type: string
          description: New admin password (leave empty to keep current)