# lib/api-zod/src/generated/types/adminSettingsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Token Gate API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminSettingsInputBackgroundStyle } from './adminSettingsInputBackgroundStyle';
import type { AdminSettingsInputPrimaryColor } from './adminSettingsInputPrimaryColor';
import type { ExamLink } from './examLink';

export interface AdminSettingsInput {
  siteName: string;
  siteDescription: string;
  primaryColor: AdminSettingsInputPrimaryColor;
  backgroundStyle: AdminSettingsInputBackgroundStyle;
  examLinks: ExamLink[];
  inactivityTimeoutSeconds: number;
  tokenWindowMinutes: number;
  useCustomToken: boolean;
  /** Fixed token value (only used when useCustomToken is true) */
  customToken?: string;
  /** New admin password (leave empty to keep current) */
  adminPassword?: string;
}