# lib/api-zod/src/generated/types/appSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Token Gate API
 * OpenAPI spec version: 0.1.0
 */
import type { AppSettingsBackgroundStyle } from './appSettingsBackgroundStyle';
import type { AppSettingsPrimaryColor } from './appSettingsPrimaryColor';
import type { ExamLink } from './examLink';

export interface AppSettings {
  siteName: string;
  siteDescription: string;
  primaryColor: AppSettingsPrimaryColor;
  backgroundStyle: AppSettingsBackgroundStyle;
  examLinks: ExamLink[];
  /** Seconds of inactivity before redirect */
  inactivityTimeoutSeconds: number;
  /** How often the token rotates in minutes */
  tokenWindowMinutes: number;
  /** Whether a fixed custom token is active */
  useCustomToken: boolean;
}