# lib/api-zod/src/generated/types/examLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Token Gate API
 * OpenAPI spec version: 0.1.0
 */

export interface ExamLink {
  label: string;
  description: string;
  url: string;
}